requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0

# Experiment Tracking & MLOps
mlflow>=2.9.0

# Model Serving
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Dashboard
streamlit>=1.29.0

# Explainability
shap>=0.43.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Data Validation
great-expectations>=0.18.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0

# Database (PostgreSQL)
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
